# examples/http_ws/openapi.json
{
  "openapi": "3.2.0",
  "info": { "title": "Demo API", "version": "1.0.0" },
  "servers": [
    {
      "url": "http://{host}/",
      "variables": {
        "host": {
          "default": "cc.apipost.cc:6002",
          "enum": ["cc.apipost.cc:6002"]
        }
      }
    }
  ],
  "x-postman-scripts": {
    "preRequest": "pm.environment.set('requestStartedAt', Date.now().toString());"
  },
  "paths": {
    "/users/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        }
      ],
      "get": {
        "operationId": "getUser",
        "summary": "Fetch a single user",
        "parameters": [
          {
            "name": "include",
            "in": "query",
            "required": false,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "items": { "type": "string", "enum": ["profile", "roles"] }
            }
          }
        ],
        "x-postman-scripts": {
          "test": [
            "pm.test('status is 200', function () { pm.response.to.have.status(200); });",
            "pm.test('payload has an id', function () { pm.expect(pm.response.json()).to.have.property('id'); });",
            "pm.environment.set('userName', pm.response.json().name);"
          ]
        },
        "responses": {
          "200": {
            "description": "The requested user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      }
    },
    "/orders": {
      "post": {
        "operationId": "createOrder",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["sku", "qty"],
                "properties": {
                  "sku": { "type": "string" },
                  "qty": { "type": "integer", "minimum": 1 }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Order created" } }
      }
    },
    "/chat/completions": {
      "post": {
        "operationId": "streamChat",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": { "type": "string", "default": "demo-model" },
                  "stream": { "type": "boolean", "default": true },
                  "messages": { "type": "array", "items": { "type": "object" } }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A stream of completion chunks",
            "content": {
              "text/event-stream": {
                "itemSchema": {
                  "type": "object",
                  "properties": { "data": { "type": "object" } }
                }
              }
            }
          }
        }
      }
    },
    "/realtime/{room}": {
      "parameters": [
        {
          "name": "room",
          "in": "path",
          "required": true,
          "schema": { "type": "string" }
        }
      ],
      "get": {
        "operationId": "joinRoom",
        "x-protocol": "websocket",
        "x-websocket": {
          "subprotocols": ["json.v1"],
          "keepAlive": { "intervalMs": 15000, "payload": "ping" }
        },
        "responses": {
          "101": { "description": "Switching protocols" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "required": ["id", "name"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" }
        }
      }
    }
  }
}